Move theme data, custom widgets to seperate files
